+++ apps/web/pages/auth/forgot-password/[id].tsx
@@ -12,7 +12,7 @@ import { Button, Input, TextField } from "@calcom/ui/v2";
 import { useLocale } from "@lib/hooks/useLocale";
 
 import { HeadSeo } from "@components/seo/head-seo";
-import AuthContainer from "@components/v2/ui/AuthContainer";
+import AuthContainer from "@components/ui/AuthContainer";
 
 type Props = {
   id: string;
