+++ apps/web/pages/settings/security/two-factor-auth.tsx
@@ -8,8 +8,8 @@ import Meta from "@calcom/ui/v2/core/Meta";
 import Switch from "@calcom/ui/v2/core/Switch";
 import { getLayout } from "@calcom/ui/v2/core/layouts/SettingsLayout";
 
-import DisableTwoFactorModal from "@components/v2/settings/DisableTwoFactorModal";
-import EnableTwoFactorModal from "@components/v2/settings/EnableTwoFactorModal";
+import DisableTwoFactorModal from "@components/settings/DisableTwoFactorModal";
+import EnableTwoFactorModal from "@components/settings/EnableTwoFactorModal";
 
 const TwoFactorAuthView = () => {
   const utils = trpc.useContext();
